encoding: Move DataRateRange from maccommands to types module
