+++ lorawan-encoding/src/maccommandcreator.rs
@@ -1,5 +1,5 @@
-use super::maccommands::{mac_commands_len, DataRateRange, SerializableMacCommand};
-use crate::types::{ChannelMask, DLSettings, Frequency, Redundancy};
+use super::maccommands::{mac_commands_len, SerializableMacCommand};
+use crate::types::{ChannelMask, DataRateRange, DLSettings, Frequency, Redundancy};
 
 #[derive(Debug, PartialEq)]
 #[cfg_attr(feature = "defmt-03", derive(defmt::Format))]
